chore: consolidate asset selection code
