+++ src/components/Modals/FiatRamps/FiatRampsModal.tsx
@@ -1,20 +1,19 @@
 import { Modal, ModalContent, ModalOverlay } from '@chakra-ui/modal'
 import { useMediaQuery } from '@chakra-ui/react'
 import type { AssetId } from '@shapeshiftoss/caip'
+import { FiatForm } from 'components/Modals/FiatRamps/views/FiatForm'
 import { useModal } from 'hooks/useModal/useModal'
 import { breakpoints } from 'theme/theme'
 
 import type { FiatRampAction } from './FiatRampsCommon'
-import { FiatManager } from './views/Manager'
 
 type FiatRampsModalProps = {
-  assetId?: AssetId
-  fiatRampAction?: FiatRampAction
+  assetId: AssetId
+  fiatRampAction: FiatRampAction
 }
 
-export const FiatRampsModal: React.FC<FiatRampsModalProps> = () => {
+export const FiatRampsModal: React.FC<FiatRampsModalProps> = props => {
   const { fiatRamps } = useModal()
-  const props = fiatRamps.props
   const { close, isOpen } = fiatRamps
   const [isLargerThanMd] = useMediaQuery(`(min-width: ${breakpoints['md']})`, { ssr: false })
 
@@ -34,10 +33,7 @@ export const FiatRampsModal: React.FC<FiatRampsModalProps> = () => {
         minWidth={{ base: '100%', md: '500px' }}
         maxWidth={{ base: 'full', md: '500px' }}
       >
-        <FiatManager
-          defaultFiatRampAction={props?.fiatRampAction}
-          defaultAssetId={props?.assetId}
-        />
+        <FiatForm assetId={props.assetId} fiatRampAction={props.fiatRampAction} />
       </ModalContent>
     </Modal>
   )
