+++ src/context/ModalProvider/ModalProvider.tsx
@@ -5,7 +5,7 @@ import React, { useMemo, useReducer } from 'react'
 import { WipeModal } from 'components/Layout/Header/NavBar/KeepKey/Modals/Wipe'
 import { BackupPassphraseModal } from 'components/Layout/Header/NavBar/Native/BackupPassphraseModal/BackupPassphraseModal'
 import { AssetSearchModal } from 'components/Modals/AssetSearch/AssetSearch'
-import { FiatRampsModal } from 'components/Modals/FiatRamps/FiatRamps'
+import { FiatRampsModal } from 'components/Modals/FiatRamps/FiatRampsModal'
 import { MobileWelcomeModal } from 'components/Modals/MobileWelcome/MobileWelcomeModal'
 import { ReceiveModal } from 'components/Modals/Receive/Receive'
 import { SendModal } from 'components/Modals/Send/Send'
